created routers folder, moved post requests for login and register into routers
