+++ server.js
@@ -7,47 +7,20 @@ const Protected = require('./middleware/protected');
 const jwt = require('jsonwebtoken');
 const secrets = require('./config/secrets');
 
-const server = express();
+const LoginRouter = require('./routers/login-router');
+const RegisterRouter = require('./routers/register-router');
 
+const server = express();
 
 server.use(helmet());
 server.use(express.json());
 server.use(cors());
 
+server.use('/api/register', RegisterRouter);
+server.use('/api/login', LoginRouter);
+
 server.get('/', (req, res) =>{
   res.send("We're connected!")
-});
-
-server.post('/api/register', (req, res) =>{
-  const user = req.body;
-  const hash = bcrypt.hashSync(user.password, 10)
-  user.password = hash;
-  Users.add(user)
-    .then(user =>{
-      res.status(200).json(user)
-    })
-    .catch(err =>{
-      res.status(500).json({message: "Could not add user"})
-    })
-})
-
-server.post('/api/login', (req, res) =>{
-  let {username, password} = req.body;
-    Users.findBy({username})
-      .first()
-      .then(user =>{
-        console.log('User info',user);
-        if (user && bcrypt.compareSync(password, user.password)){
-          const token = generateToken(user);
-          console.log('token', token);
-          res.status(200).json({message: `${user.username} is logged in!`,token,})
-        } else {
-          res.status(400).json({message: "Invalid username or password"})
-        }
-      })
-      .catch(err =>{
-        res.status(500).json(err)
-      })
 });
 
 server.get('/api/users', Protected, (req, res) =>{
@@ -60,16 +33,6 @@ server.get('/api/users', Protected, (req, res) =>{
   })
 });
 
-function generateToken(user){
-  const payload ={
-    username: user.username,
-    subject: user.id,
-    department: user.department
-  };
-  const options ={
-    expiresIn: "8h"
-  };
-  return jwt.sign(payload, secrets.jwtSecret, options);
-}
+
 
 module.exports = server;
